fixed cors routing path and cleaned up file structure
