+++ web-scrapper-back-end/flaskr/controller/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+loginBluePrint = Blueprint('loginBluePrint', __name__, static_folder='controller.login')
+
+dataAggregationBluePrint = Blueprint('dataAggregationBluePrint', __name__, static_folder='controller.dataAggregation')
