+++ web-scrapper-front-end/package.json
@@ -37,6 +37,5 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  },
-  "proxy":"http://localhost:8080"
+  }
 }
