--- webScrapperBackEnd/setup.py
@@ -1,10 +1,0 @@
-from setuptools import setup
-
-setup(
-    name='webScrapperBackEnd',
-    packages=['webScrapperBackEnd'],
-    include_package_data=True,
-    install_requires=[
-        'flask',
-    ],
-)
